langchain-chatbot/requirements.txt
langchain==0.1.0
streamlit==1.28.1
openai==1.3.7
python-dotenv==1.0.0
langchain-openai==0.0.2
langchain-community==0.0.10